+++ src/modules/command/command.module.ts
@@ -8,16 +8,16 @@ import { BugReport } from './commands/bug';
 import { CoinflipCommand } from './commands/coinflip';
 import { GoldCommand } from './commands/gold';
 import SomeoneOnceSaidCommand from './commands/someone-once-said';
-import { SomeoneOnceSaidModule } from '../someone-once-said/module/someone-once-said.module';
 import GetRandomQuote from './commands/get-a-quote';
 import { PollCommand } from './commands/poll';
-import { PollModule } from '../poll/module/poll.module';
-import { VersionModule } from '../version/module/version.module';
 import { VersionCommand } from './commands/version';
-import { BirthdayEntryModule } from '../birthday/module/birthday-entry.module';
 import AddBirthdayEntryCommand from './commands/add-birthday-entry';
 import ActivateBirthdayEntryShoutoutCommand from './commands/activate-birthday-shoutout';
 import DeactivateBirthdayEntryShoutoutCommand from './commands/deactivate-birthday-shoutout';
+import { SomeoneOnceSaidModule } from '../models/someone-once-said/module/someone-once-said.module';
+import { PollModule } from '../models/poll/module/poll.module';
+import { VersionModule } from '../models/version/module/version.module';
+import { BirthdayEntryModule } from '../models/birthday/module/birthday-entry.module';
 
 @Module({
   providers: [
