+++ src/modules/command/commands/deactivate-birthday-shoutout.ts
@@ -1,8 +1,8 @@
 import { CacheType, CommandInteraction, SlashCommandBuilder } from 'discord.js';
 import { ACommand } from '../command.abstract';
 import { Inject } from '@nestjs/common';
-import { BirthdayEntryService } from '../../birthday/service/birthday-entry.service';
-import { CreateOrUpdateBirthdayEntryDto } from '../../birthday/dto/create-or-update-birthday-entry.dto';
+import { BirthdayEntryService } from '../../models/birthday/service/birthday-entry.service';
+import { CreateOrUpdateBirthdayEntryDto } from '../../models/birthday/dto/create-or-update-birthday-entry.dto';
 
 export default class DeactivateBirthdayEntryShoutoutCommand extends ACommand {
   constructor(
