+++ src/modules/cron-tasks/task.module.ts
@@ -1,5 +1,5 @@
 import { Module } from "@nestjs/common";
-import { BirthdayEntryModule } from "../birthday/module/birthday-entry.module";
+import { BirthdayEntryModule } from "../models/birthday/module/birthday-entry.module";
 
 
 @Module({
