+++ src/modules/discord/discord.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { DiscordService } from './discord.service';
 import { AppConfigModule } from '../../config/config.module';
-import { BirthdayEntryModule } from '../birthday/module/birthday-entry.module';
+import { BirthdayEntryModule } from '../models/birthday/module/birthday-entry.module';
 
 @Module({
   imports: [AppConfigModule, BirthdayEntryModule],
