+++ src/modules/event/services/interaction.spec.ts
@@ -1,7 +1,7 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { Interaction } from './interaction';
 import { CommandService } from '../../command/command.service';
-import { PollService } from '../../poll/service/poll.service';
+import { PollService } from '../../models/poll/service/poll.service';
 
 describe('Interaction', () => {
   let service: Interaction;
