+++ src/modules/models/birthday/module/birthday-entry.module.ts
@@ -1,8 +1,8 @@
 import { Module } from '@nestjs/common';
 import { MongooseModule } from '@nestjs/mongoose';
-import { BirthdayEntry, BirthdayEntrySchema } from '../../../schemas/birthday-entry.schema';
 import { BirthdayEntryService } from '../service/birthday-entry.service';
 import { BirthdayEntryController } from '../controller/birthday-entry.controller';
+import { BirthdayEntry, BirthdayEntrySchema } from '../../../../schemas/birthday-entry.schema';
 
 @Module({
   imports: [
