+++ src/modules/models/birthday/service/birthday-entry.service.ts
@@ -1,7 +1,7 @@
 import { InjectModel } from '@nestjs/mongoose';
 import { Model } from 'mongoose';
-import { BirthdayEntryDocument } from '../../../schemas/birthday-entry.schema';
 import { CreateOrUpdateBirthdayEntryDto } from '../dto/create-or-update-birthday-entry.dto';
+import { BirthdayEntryDocument } from '../../../../schemas/birthday-entry.schema';
 
 export class BirthdayEntryService {
   constructor(
