+++ src/modules/models/poll/module/poll.module.ts
@@ -1,8 +1,8 @@
 import { Module } from '@nestjs/common';
 import { MongooseModule } from '@nestjs/mongoose';
-import { Poll, PollSchema } from '../../../schemas/poll.schema';
 import { DbPollService } from '../service/db-poll.service';
 import { PollService } from '../service/poll.service';
+import { Poll, PollSchema } from '../../../../schemas/poll.schema';
 
 @Module({
   imports: [
