+++ src/modules/models/someone-once-said/module/someone-once-said.module.ts
@@ -1,10 +1,7 @@
 import { Module } from '@nestjs/common';
 import { MongooseModule } from '@nestjs/mongoose';
 import { SomeoneOnceSaidService } from '../service/someone-once-said.service';
-import {
-  SomeoneOnceSaid,
-  SomeoneOnceSaidSchema,
-} from '../../../schemas/someone-once-said.schema';
+import { SomeoneOnceSaid, SomeoneOnceSaidSchema } from '../../../../schemas/someone-once-said.schema';
 
 @Module({
   imports: [
