+++ src/modules/models/someone-once-said/service/someone-once-said.service.spec.ts
@@ -1,9 +1,9 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { Model, Error, MongooseError } from 'mongoose';
 import { SomeoneOnceSaidService } from './someone-once-said.service';
-import { SomeoneOnceSaidEntity } from '../../../schemas/someone-once-said-entity.model';
-import { SomeoneOnceSaidDocument } from '../../../schemas/someone-once-said.schema';
 import { getModelToken } from '@nestjs/mongoose';
+import { SomeoneOnceSaidDocument } from '../../../../schemas/someone-once-said.schema';
+import { SomeoneOnceSaidEntity } from '../../../../schemas/someone-once-said-entity.model';
 describe('SomeoneOnceSaidService', () => {
   let service: SomeoneOnceSaidService;
   let modelMock: Model<SomeoneOnceSaidDocument>;
